Cleaned transcode decision tracker
