+++ src/libs/subsonic/CMakeLists.txt
@@ -2,6 +2,7 @@
 add_library(lmssubsonic STATIC
 	impl/endpoints/transcoding/AudioFileInfo.cpp
 	impl/endpoints/transcoding/TranscodeDecision.cpp
+	impl/endpoints/transcoding/TranscodeDecisionTracker.cpp
 	impl/endpoints/AlbumSongLists.cpp
 	impl/endpoints/Bookmarks.cpp
 	impl/endpoints/Browsing.cpp
